requirements.txt
# Core Data Processing
pandas>=2.1
numpy>=1.26
duckdb>=0.9

# dbt
dbt-core>=1.7
dbt-duckdb>=1.7

# Workflow Orchestration (optional - for production)
# prefect>=2.14

# Data Quality
great-expectations>=0.18

# Network Analysis
networkx>=3.2

# Natural Language Processing
sentence-transformers>=2.2
transformers>=4.35
datasets>=2.16
# torch>=2.1  # Installed separately in Dockerfile with CUDA support

# Machine Learning
scikit-learn>=1.3
umap-learn>=0.5.6

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Object Storage
minio>=7.2
boto3>=1.34

# Configuration
python-dotenv>=1.0

# Utilities
tqdm>=4.66
pyarrow>=14.0
joblib>=1.3  # For parallel processing

# Testing
pytest>=7.4
pytest-cov>=4.1

# Code Quality
black>=23.12
flake8>=7.0
isort>=5.13

# Jupyter (optional, for exploration)
jupyter>=1.0
ipykernel>=6.27
